Update README.md
In this update, I rewrote the part where R was used to adapt the OTU and tax table, to be able to perform it directly in the server with Python. Therefore, it is no longer necessary to download and re-upload files from and to the server. 
Besides, the code to write the manifest file as preparation for DADA2 was rewritten slightly to improve re-usability. 
Furthermore, some minor changes were done to improve legibility of the pipeline.